feat: rename file to profileRouter
