+++ package.json
@@ -40,7 +40,7 @@
     "cmdk": "^0.2.1",
     "date-fns": "^3.6.0",
     "exceljs": "^4.4.0",
-    "expo-backend-types": "^0.9.0",
+    "expo-backend-types": "^0.10.0",
     "fast-csv": "^5.0.1",
     "firebase": "^10.13.0",
     "firebase-admin": "^12.3.1",
